update:move SwitchAreaTargetInfo in it's own header
